NOA Group Assigner
Easy way to assign outfit from all outfits with same output to one
of three groups (nude, outfit, armor).

Shows list of all outfits that would be built with batch build with
the ability to choose one outfit for each output file and option
to automatically assign that outfit to one of three groups.

Supports preview of outfits, save & load of groups to/from single
group xml file. User can also pick preset for each of the groups to
be used in the preview.